requirements.txt

discord.py>=2.3.0
google-generativeai>=0.3.0
flask>=2.3.0
psutil>=5.9.0
asyncio
aiohttp>=3.8.0
python-dotenv>=1.0.0
replit
